circle royale v.1

+++ shared/src/main/scala/be/tzbob/circleroyale/Camera.scala
@@ -1,6 +1,6 @@
 package be.tzbob.circleroyale
 
-case class Camera(position: Vec2D, width: Int, height: Int) {
+case class Camera(position: Vec2D, width: Double, height: Double) {
   val viewboxText: String = {
     val Vec2D(x, y) = position
     s"$x $y $width $height"
